pyproject.toml
[project]
name = "punishbot"
version = "0.1.0"
description = "A bot just for punishment to 𝒕𝒉𝒆 𝑬𝒎𝒑𝒆𝒓𝒐𝒓’𝒔 𝒆𝒏𝒆𝒎𝒚."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.1",
    "pydantic>=2.10.4",
    "pydantic-settings>=2.7.0",
    "discord-py>=2.4.0",
]

[dependency-groups]
dev = [
    "black>=24.10.0",
    "flake8>=7.1.1",
    "isort>=5.13.2",
    "mypy>=1.13.0",
    "pytest>=8.3.4",
]

[tool.mypy]
strict = true
exclude = ["^playground/"]

[tool.isort]
profile = "black"
skip_gitignore = true
skip = ["^playground/"]
